Fix link error on test_resampler build.
